Readabillity
Just did a little makeover of the most button and form names.
To make it a lot more readable and understandable if you want to Read the sample
