+++ CmlLibSample/Program.cs
@@ -16,7 +16,7 @@ namespace CmlLibSample
         {
             Application.EnableVisualStyles();
             Application.SetCompatibleTextRenderingDefault(false);
-            Application.Run(new Form1());
+            Application.Run(new Main());
         }
     }
 }
